Generation Context Cleanup and other fixes  (#74)
* refactor: create decorator for generation config

Adds a decorator function which sets up the generation context for group
commands. This contains the standard paramters which control generation
across all of our commands which call the template backend

* refactor(query)!: refresh_output replacement

replaces refresh_output with use_cache to highlight what this does to
users. Also flips the behaviour based on the name of the variable
I.e. where refresh_output = False - use_cache = True

* refactor: update commands to align with refactor

Updates the commands to support the pass_generation and use_cache
changes in the previous commits

* refactor: update tests for cache change

* refactor: remove unused test deploy

* refactor: remove unused context

* refactor(query): create blueprint context

Adds a new context Class BlueprintContext which extends the Generation
Context. This provides additional information to handle the generation
of blueprint and passing extra information when running queries

* fix: handle failing tests in command

A fix to the test command which was ignoring failed results from
pytest. Now failing tests are treated and handled as an exception

* fix: remove old generation context

* chore: linting fixes