feat: multiple diagram generation
Adds support for generating multiple diagrams from a single command and
aligns with the user generated diagram approach. This uses the same
approach as the deployments commands which query the available commands
and apply regex filtering based on the provided id

Adds a list diagrams command to show the available commands

Splits the output directories for source code and the generated diagram
images. This allows for users to decide on what they require and where
they can store each output type