refactor!: options context and generation parameters (#89)
* refactor!: config context handling

This is a major refactor of the configuration context that we use to
define standard configration applied to hamlet commands

- The generation framework, provider and input source are now top level
commands and are part of the Options configuration class
- Environment variable paramters are now provided to each runner call
rather than having env vars set at the start of the command. This shows
where these parameters are being used and allows them to be accessed by
the python executor
- Added log-level as a configuration option so that it can be added on
the command line. It also respects the existing GENERATION_LOG_LEVEL
setting and will prefer that
- cleans up paramter passing to remove touch points

* fix: linting updates

* feat: add env options to setup

* fix: add generation options back

* feat: support for options on run

* refactor: update log configuration options