fix: handle large outputs in runner calls
Some errors that are raised from hamlet include full blueprints or large
exceptions. The subprocess.wait() function can deadlock on large outputs
when the allocated buffer is larger than the message.

This moves to using communicate() which is the recommended approach when
dealing with outputs

Also moved to using which to locate bash for compatability with
different installations and confirm that generation dir is available
before running the command