feat: add component group
Adds the component command group which is used to query and discover
information about components and occurrences.

The two initial commands describe-occurrences and list-occurrences
are similar to the query command group but use a new data format
available from the engine to simplify processing

The new format also handles suboccurrences and uses the name of the
occurrence as the key for search queries to reduce the number of
paramters required to use the command. This also removes confusion
around instances and versions which are optional in the hamlet engine