fix: logging during cli execptions
adds a fix for the command runner to ensure that we collect and report
outputs on exceptions. The stdout and stderr are split as its difficult
to merge them. But this at least makes output visible

fix for networkx install reqiurements which were breaking with changes
between cfn-lint and networkx