feat: contract based deployments
adds support in the cli for running deployments
using contracts instead of bash scripts

This allows for the engine to control how the
deployment is run and also lets the python
executor look after making it happen

Current implementation handles cloudformation
deployments using either Change Sets or standard
stack updates. Using standard stack updates should
reduce some of the processing time when running
deployments that don't need change sets.

Also supports presigned url hosting of
cloudformation templates so that we can use
larger template sizes

To make the bites to chomp off smaller there is a
utility function which calls the bash executor
to return the cf_dir so that we don't have to
fully replicate the context lookup to get this
going